REF: app-storage class
