+++ class/index.ts
@@ -18,3 +18,4 @@ export * from './wallets/multisig-hd-wallet';
 export * from './wallets/slip39-wallets';
 export * from './hd-segwit-bech32-transaction';
 export * from './multisig-cosigner';
+export * from './app-storage';
